+++ cdp-agentkit-core/python/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+- Add action `superfluid_create_flow` to create a flow using Superfluid
+- Add action `superfluid_update_flow` to update a flow using Superfluid
+- Add action `superfluid_delete_flow` to delete a flow using Superfluid
+
 ## [0.0.10] - 2025-01-22
 
 ### Added